Add files via upload
First commitment. I uploaded an example of temperature recording to calculate the time constant of a system (EFI234105840_20230801164648.xls) and the first part of the script for data fitting. This script just takes the excel file and the fitting interval as inputs from command line and plots them in a graph, so that the user can check if this first step is performed correctly.